customer-churn-prediction/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
matplotlib==3.7.2
seaborn==0.12.2
shap==0.42.1
imbalanced-learn==0.11.0
jupyter==1.0.0
mysql-connector-python==8.1.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
plotly==5.15.0
kaleido==0.2.1